refactor(about): move motifs to shared; add Motifs to About Goal component; adjust for small screen
